requirements.txt
langchain
openai
python-dotenv
numpy
streamlit
pandas